regression analysis of arts calibration
