+++ flights.csv
@@ -1,0 +1,33 @@
+,flightletter,location
+20240809,a,transfer
+20240811,a,sal
+20240813,a,sal
+20240816,a,sal
+20240818,a,sal
+20240821,a,sal
+20240822,a,sal
+20240825,a,sal
+20240827,a,sal
+20240829,a,sal
+20240831,a,sal
+20240903,a,sal
+20240906,a,transfer
+20240907,a,barbados
+20240909,a,barbados
+20240912,a,barbados
+20240914,a,barbados
+20240916,a,barbados
+20240919,a,barbados
+20240921,a,barbados
+20240923,a,barbados
+20240924,a,barbados
+20240926,a,barbados
+20240928,a,barbados
+20240929,a,transfer
+20241105,a,op
+20241107,a,op
+20241110,a,op
+20241112,b,op
+20241114,b,op
+20241116,a,op
+20241119,a,op
